refactor: dom utils
